refactor(board-service): allow passing multiple properties of an object instead a single key/value pair
Like the team discussed, we are changing the updateBoard function to
accept an updates object that can contain multiple properties to be
updated at once.

This also remove temporarily all logic related to activities until we
we specify how they should be implemented.

I also changed the method to just updateBoard to be more consistent with
its purpose, as well as, broke down the updates into smaller and more
specific function for future implementation of individual updates of
board/lists/cards by the server.